Mover calcularNúmeroPasajeros() a etl.Añadir
